+++ Tests/Integration/MSFT_xBLBitlocker.config.ps1
@@ -20,7 +20,7 @@ else
 
 <#
     .SYNOPSIS
-        Enables Bitlocker on the Operating System drive using a TpmProtector
+        Enables Bitlocker on the Operating System drive using a TpmProtector.
 #>
 Configuration MSFT_xBLBitlocker_BasicTPMEncryptionOnSysDrive_Config
 {
@@ -32,6 +32,29 @@ Configuration MSFT_xBLBitlocker_BasicTPMEncryptionOnSysDrive_Config
         {
             MountPoint       = $env:SystemDrive
             PrimaryProtector = 'TpmProtector'
+            UsedSpaceOnly    = $true
+        }
+    }
+}
+
+<#
+    .SYNOPSIS
+        Enables Bitlocker on the Operating System drive using a TpmProtector
+        and passed multiple Switch parameters of Enable-Bitlocker with False
+        values.
+#>
+Configuration MSFT_xBLBitlocker_TPMEncryptionOnSysDriveWithFalseSwitchParams_Config
+{
+    Import-DscResource -ModuleName 'xBitlocker'
+
+    Node $AllNodes.NodeName
+    {
+        xBLBitlocker Integration_Test
+        {
+            MountPoint         = $env:SystemDrive
+            PrimaryProtector   = 'TpmProtector'
+            HardwareEncryption = $false
+            UsedSpaceOnly      = $false
         }
     }
 }